Move the merging code from deduplicate code to common module
That way it can be used in a new command to merge individual items.
